remove some alias file
