+++ src/main/scala/vsys/Application.scala
@@ -102,8 +102,6 @@ class Application(val actorSystem: ActorSystem, val settings: VsysSettings) exte
       //AssetsBroadcastApiRoute(settings.restAPISettings, utxStorage, allChannels),
       LeaseApiRoute(settings.restAPISettings, wallet, utxStorage, allChannels, stateReader, time),
       LeaseBroadcastApiRoute(settings.restAPISettings, utxStorage, allChannels),
-      //AliasApiRoute(settings.restAPISettings, wallet, utxStorage, allChannels, time, stateReader),
-      //AliasBroadcastApiRoute(settings.restAPISettings, utxStorage, allChannels),
       SPOSApiRoute(settings.restAPISettings, wallet, utxStorage, allChannels, time, stateReader),
       SPOSBroadcastApiRoute(settings.restAPISettings, utxStorage, allChannels),
       ContractApiRoute(settings.restAPISettings, wallet, utxStorage, allChannels, time, stateReader),
@@ -127,8 +125,6 @@ class Application(val actorSystem: ActorSystem, val settings: VsysSettings) exte
       //classOf[AssetsBroadcastApiRoute],
       classOf[LeaseApiRoute],
       classOf[LeaseBroadcastApiRoute],
-      //classOf[AliasApiRoute],
-      //classOf[AliasBroadcastApiRoute],
       classOf[SPOSApiRoute],
       classOf[SPOSBroadcastApiRoute],
       classOf[ContractApiRoute],
