--- src/main/scala/vsys/api/http/alias/CreateAliasRequest.scala
@@ -1,15 +1,0 @@
-package vsys.api.http.alias
-
-import io.swagger.annotations.ApiModelProperty
-import play.api.libs.json.{Format, Json}
-
-case class CreateAliasRequest(@ApiModelProperty(value = "Base58 encoded sender public key", required = true)
-                              sender: String,
-                              @ApiModelProperty(value = "Alias", required = true)
-                              alias: String,
-                              @ApiModelProperty(required = true)
-                              fee: Long)
-
-object CreateAliasRequest {
-  implicit val aliasRequestFormat: Format[CreateAliasRequest] = Json.format
-}
