+++ src/main/scala/vsys/blockchain/transaction/lease/LeaseTransaction.scala
@@ -3,7 +3,7 @@ package vsys.blockchain.transaction.lease
 import com.google.common.primitives.{Bytes, Longs, Shorts}
 import vsys.blockchain.state.ByteStr
 import play.api.libs.json.{JsObject, Json}
-import vsys.account.{Address, AddressOrAlias, PrivateKeyAccount, PublicKeyAccount}
+import vsys.account.{Address, PrivateKeyAccount, PublicKeyAccount}
 import vsys.blockchain.transaction.TransactionParser._
 import vsys.blockchain.transaction._
 import vsys.blockchain.transaction.ProvenTransaction
@@ -15,7 +15,7 @@ case class LeaseTransaction private(amount: Long,
                                     fee: Long,
                                     feeScale: Short,
                                     timestamp: Long,
-                                    recipient: AddressOrAlias,
+                                    recipient: Address,
                                     proofs: Proofs)
   extends ProvenTransaction {
 
@@ -45,7 +45,7 @@ object LeaseTransaction {
 
   def parseTail(bytes: Array[Byte]): Try[LeaseTransaction] = Try {
     (for {
-      recRes <- AddressOrAlias.fromBytes(bytes, 0)
+      recRes <- Address.fromBytes(bytes, 0)
       (recipient, recipientEnd) = recRes
       quantityStart = recipientEnd
       quantity = Longs.fromByteArray(bytes.slice(quantityStart, quantityStart + 8))
@@ -62,7 +62,7 @@ object LeaseTransaction {
              fee: Long,
              feeScale: Short,
              timestamp: Long,
-             recipient: AddressOrAlias,
+             recipient: Address,
              proofs: Proofs): Either[ValidationError, LeaseTransaction] = {
 
     if (amount <= 0) {
@@ -87,7 +87,7 @@ object LeaseTransaction {
              fee: Long,
              feeScale: Short,
              timestamp: Long,
-             recipient: AddressOrAlias): Either[ValidationError, LeaseTransaction] = for {
+             recipient: Address): Either[ValidationError, LeaseTransaction] = for {
     unsigned <- createWithProof(amount, fee, feeScale, timestamp, recipient, Proofs.empty)
     proofs <- Proofs.create(List(EllipticCurve25519Proof.createProof(unsigned.toSign, sender).bytes))
     tx <- createWithProof(amount, fee, feeScale, timestamp, recipient, proofs)
@@ -98,7 +98,7 @@ object LeaseTransaction {
              fee: Long,
              feeScale: Short,
              timeStamp: Long,
-             recipient: AddressOrAlias,
+             recipient: Address,
              signature: ByteStr): Either[ValidationError, LeaseTransaction] = for {
     proofs <- Proofs.create(List(EllipticCurve25519Proof.buildProof(sender, signature).bytes))
     tx <- createWithProof(amount, fee, feeScale, timeStamp,recipient, proofs)
