--- src/test/scala/vsys/account/AddressOrAliasSpecification.scala
@@ -1,32 +1,0 @@
-package vsys.account
-
-import org.scalatest.prop.{GeneratorDrivenPropertyChecks, PropertyChecks}
-import org.scalatest.{Matchers, PropSpec}
-import vsys.blockchain.transaction.TransactionGen
-
-class AddressOrAliasSpecification extends PropSpec with PropertyChecks with GeneratorDrivenPropertyChecks with Matchers with TransactionGen {
-
-  property("Account serialization round trip") {
-    forAll(accountGen) { account: PrivateKeyAccount =>
-      val bytes = account.bytes.arr
-      val address = Address.fromBytes(bytes).right.get
-      address.stringRepr shouldBe account.stringRepr
-    }
-  }
-
-  property("Alias serialization round trip") {
-    forAll(aliasGen) { alias: Alias =>
-      val bytes = alias.bytes.arr
-      val representation = Alias.fromBytes(bytes).right.get
-      representation.stringRepr shouldBe representation.stringRepr
-    }
-  }
-
-  property("AccountOrAlias serialization round trip") {
-    forAll(accountOrAliasGen) { aoa: AddressOrAlias =>
-      val bytes = aoa.bytes.arr
-      val addressOrAlias = AddressOrAlias.fromBytes(bytes, 0).right.get
-      addressOrAlias._1.stringRepr shouldBe aoa.stringRepr
-    }
-  }
-}
